+++ src/Shared/Infrastructure/LLMChain/SettingsToolBox.php
@@ -7,23 +7,38 @@ namespace ChronicleKeeper\Shared\Infrastructure\LLMChain;
 use ChronicleKeeper\Settings\Application\SettingsHandler;
 use PhpLlm\LlmChain\Response\ToolCall;
 use PhpLlm\LlmChain\ToolBox\Metadata;
-use PhpLlm\LlmChain\ToolBox\ToolBox;
 use PhpLlm\LlmChain\ToolBox\ToolBoxInterface;
+
+use function array_key_exists;
 
 class SettingsToolBox implements ToolBoxInterface
 {
     public function __construct(
         private readonly SettingsHandler $settingsHandler,
-        private readonly ToolBox $llmToolBox,
+        private readonly ToolBoxInterface $llmToolBox,
     ) {
     }
 
     /** @return Metadata[] */
     public function getMap(): array
     {
-        $toolMap = $this->llmToolBox->getMap();
-
+        $descriptions = $this->settingsHandler->get()->getChatbotFunctions()->getFunctionDescriptions();
+        $toolMap      = $this->llmToolBox->getMap();
         // Check Settings for custom tool names and descriptions and overhault the metadata here :)
+
+        foreach ($toolMap as $index => $tool) {
+            if (! array_key_exists($tool->name, $descriptions) || $descriptions[$tool->name] === '') {
+                continue;
+            }
+
+            $toolMap[$index] = new Metadata(
+                $tool->className,
+                $tool->name,
+                $descriptions[$tool->name],
+                $tool->method,
+                $tool->parameters,
+            );
+        }
 
         return $toolMap;
     }