+++ tests/Shared/Infrastructure/LLMChain/Stub/ExampleTool.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ChronicleKeeper\Test\Shared\Infrastructure\LLMChain\Stub;
+
+use PhpLlm\LlmChain\ToolBox\Attribute\AsTool;
+
+#[AsTool('ExampleTool', 'This is an example tool.')]
+class ExampleTool
+{
+    /** @param string $foo This is an example description. */
+    public function __invoke(string $foo): string
+    {
+        return $foo;
+    }
+}